Move http package into protocols
